@@ -1,19 +1,19 @@
 name: enough_mail_html
-description: Generate HTML from an email mime message. Eases handling of `MimeMessage` in the `enough_mail` package.
+description: Generate HTML from an email mime message. Eases handling of `MimeMessage` in the [enough_mail](https://pub.dev/packages/enough_mail) package.
 version: 0.4.0
 homepage: https://github.com/Enough-Software/enough_mail_html
 
 environment:
-  sdk: '>=2.9.3 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   html: ^0.15.0
-  enough_mail: #'>=0.3.0 <2.0.0'
+  enough_mail: #'>=1.0.0 <2.0.0'
     git:
       url: https://github.com/Enough-Software/enough_mail.git
-    #path: ../enough_mail/
-  image: '>=2.0.0 <4.0.0'
+    # path: ../enough_mail/
+  image: ^3.0.1
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  pedantic: ^1.11.0
+  test: ^1.16.7
