@@ -32,7 +32,7 @@ void main() {
 
       final meta = Element.html(
           '<meta name="viewport" content="width=device-width, initial-scale=1.0">');
-      document.head.append(meta);
+      document.head!.append(meta);
       final imageElements = document.getElementsByTagName('img');
       for (final imageElement in imageElements) {
         final src = imageElement.attributes['src'];
