@@ -1,3 +1,5 @@
+import 'package:enough_mail/enough_mail.dart';
+import 'package:enough_mail_html/enough_mail_html.dart';
 import 'package:enough_mail_html/src/text/linebreak_text_transformer.dart';
 import 'package:test/test.dart';
 
@@ -6,7 +8,10 @@ void main() {
     test('linebreaks', () {
       final input = 'hello\r\nworld.\n';
       final transformer = LineBreakTextTransformer();
-      expect(transformer.transform(input, null, null), 'hello<br/>world.<br/>');
+      expect(
+          transformer.transform(input, MimeMessage(),
+              TransformConfiguration.standardConfiguration),
+          'hello<br/>world.<br/>');
     });
   });
 }
